orca/pyproject.toml
[project]
name = "orca"
version = "2.0.0a1"
description = "Multi-Agentic Platform for Binary Analysis, Malware Triage & Network Protocol Analysis"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "pydantic>=2.0",
    "langgraph>=0.2",
    "langchain-core>=0.3",
    "langchain-openai>=0.2",
    "litellm>=1.40",
    "pyhidra>=1.0",
]

[project.optional-dependencies]
binja = []  # Binary Ninja is proprietary — user installs separately
network = ["scapy>=2.5"]
yara = ["yara-python>=4.3"]
sandbox = ["paramiko>=3.0"]
embeddings = ["numpy>=1.24", "sentence-transformers>=2.2"]
all = ["scapy>=2.5", "yara-python>=4.3", "paramiko>=3.0", "numpy>=1.24"]

[project.scripts]
orca = "orca.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
